Remove sentence because it is confusing
"We hope to be able to offer 40-50 pre-college scholarships per year by 2012." This sentence doesn't make much sense because this whole article is written as a reflextion piece. It would be more valuabe to talk about what you accomplished or rewrite it so that is is clear that his was a goal in 2007/2008. I think the section is stronger focusing on what happened, but it is possible to include this objective if you find it valuable.