Cleanup featured article image
Because this is an image we should not use a header element. It is not semantic HTML. I renamed this section to image because a banner is an element that typically spans the full width of a section and is used to highlight/announce information.

I added the aria role to signify to screen readers that this section is purely presentational, i.e. it doesn't have any content of value. It is just for visual presentation.

I removed the class `featurd-article__content` because it is not needed. Flex applies at the parent level.

In the future, I think adding an icon for PDF or paper would make it look much nicer.