Cleanup CSS and design
We can achieve a similar CSS effect with less code. There was also a lot of mixing of code and properties.

I aimemed to align the content horizontally with the grid above. Will need to adjust the article grid above later. It also is using a lot of extra code.